waf: strict posix c libc++ (#3)
Add a compiled llvm's libc++ to get rid of the gnu extensions enforced in gnu's
libstdc++. This way, the resulting link will depend on the same libraries as Go's:
libc.so, libpthread.so and the elf loader.

It allows to keep the strong Go property of getting from the compilation a static
program that can be executed on any linux platform. So for example, the copy
from a debian development environment (glibc) to an alpine production
environment (muslc) should work.