waf: export waf error values
Make error values public so that the agent can further manage them. In this
case, we need to check the timeout error to avoid sending an error.

Note that they are redefined in types.go in order to don't "leak" the C types:
as we want to be able to compile without cgo, the bindings need to convert to
the redefined Go types.