waf: avoid string copies in go bindings when running a rule
Remove the string copy performed by `C.CString()` when creating a PWArgs string.
The PWArgs string constructor function will copy it into a new C char buffer. So
it is fine not to copy it during the time of the call.

This is a major source of performance improvement which allows to run much
faster, as most of the WAF inputs are strings.